README.md
The above "War" card game file contains 4 suits and all 13 standard card ranks. This program implements the game in the following steps:
1.) Declare 52 cards (a standard deck).
2.) Give 26 unique cards to the player engaged in the program.
3.) Randomize the other 26 cards to be stored into the computer.
4.) Shuffle the cards of the user.
5.) Shuffle the cards of the computer.
6.) Show the cards of both the player and the computer.
7.) Distribute the points and erase the cards.
8.) Pause the game, and then press any key to continue.
9.) Conclude the session with a winner!
The file contains basic C++ programming for the most part, but it includes abstraction data types such as classes that allow the program to initialize the card objects in the process of the game.